+++ CHANGELOG.md
@@ -525,3 +525,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix too many characters going offscreen on the custom mode screen (kiooeht)
 
 #### dev ####
+* Fix power console command for some powers (admiralbolt)
+* Fix power console command not stacking powers correctly (admiralbolt)