+++ CHANGELOG.md
@@ -530,3 +530,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix whatmod not working on linked relics (kiooeht)
 * Add option for Custom Mode button (JohnnyDevo)
 * Fix custom character unlocks (MichaelMayhem)
+* Fix power console command for some powers (admiralbolt)
+* Fix power console command not stacking powers correctly (admiralbolt)